requirements.txt
amqp==2.2.2
Babel==2.1.1
billiard==3.5.0.3
blinker==1.4
celery==4.1.0
coverage==4.0.1
decorator==4.0.4
Flask==0.10.1
Flask-Babel==0.9
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
flipflop==1.0
guess-language==0.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.8
kombu==4.1.0
langid==1.1.5
MarkupSafe==0.23
numpy==1.10.1
pbr==1.8.1
pkg-resources==0.0.0
python-magic==0.4.10
python-openid==2.2.5
pytz==2015.7
redis==2.10.6
regex==2017.12.12
six==1.10.0
speaklater==1.3
SQLAlchemy==1.0.9
sqlalchemy-migrate==0.10.0
sqlparse==0.1.18
Tempita==0.5.2
vine==1.1.4
Werkzeug==0.10.4
Whoosh==2.7.0
WTForms==2.0.2
